+++ src/Components/TopNavBar.tsx
@@ -1,7 +1,7 @@
 import { Image, Link, Portal, Text, makeStyles, shorthands, tokens } from "@fluentui/react-components";
 import { Flex, NavH, NavW } from "~/Helpers/Styles";
 import { AdminTopNav } from "~/Pages/Admin/Nav";
-import { AdminProductAddButton } from "~/Pages/Admin/Product/ProductAdd";
+import { AdminProductAddButton } from "~/Pages/Admin/Product/Add";
 import { AdminUserSearch } from "~/Pages/Admin/User/Search";
 import { AvatarMenu } from "./AvatarMenu";
 import { ShopCart } from "./ShopCart";
